Split omnibar suggestions overlay types
